pyproject.toml
[project]
name = "verteilte-systeme"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fire (>=0.7.0,<0.8.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "googleapis-common-protos (>=1.70.0,<2.0.0)",
    "grpcio-status (>=1.71.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
